temp_scraper_repo/requirements.txt
# Authentication and Authorization
# Database and ORM
# Environment and configuration
# Security
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask==2.3.2
Pillow==10.0.0
SQLAlchemy==2.0.19
Werkzeug==2.3.6
beautifulsoup4==4.12.2
cryptography==41.0.3
flask-dance==7.0.0
flask-login==0.6.3
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
flask-talisman==1.1.0
flask-wtf==1.2.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
mss==9.0.1
openai>=0.27.0 
opencv-python==4.8.1.78
paypalrestsdk==1.13.1
pyodbc==4.0.39
python-dotenv==1.0.0
python-dotenv==1.0.1
requests-oauthlib==1.3.1
requests==2.31.0
